requirements.txt
discord.py==2.3.1
requests==2.31.0